Refactor FormField
